elf_parser: use section headers if DT_REL points outside program segments
At least until rust-bpf-sysroot v0.13, we used to generate invalid
dynamic sections where the address of DT_REL was not contained in any
program segment. When loading such files, fallback to relying on section
headers to find out the offset of the relocations table.

One notable program that exhibits this bug is
https://github.com/solana-labs/solana-program-library/releases/tag/memo-v3.0.0